feat: create exercise loader module
