+++ exercises/pronouns.html
@@ -57,6 +57,6 @@
 </div>
 
 <script src="../js/exercise-engine.js"></script>
-<script src="../js/pronouns-exercise.js"></script>
+<script type="module" src="../js/pronouns-exercise.js"></script>
 </body>
 </html>
